% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_avg_beta.R
\name{add_avg_beta}
\alias{add_avg_beta}
\title{Add beta-values to table with means of M-values}
\usage{
add_avg_beta(tt)
}
\arguments{
\item{tt}{Data frame with columns that have means of M-values and end with \code{.avg}.}
}
\value{
Data frame with additional columns of mean beta-values.
}
\description{
Add beta-values to table with means of M-values, where beta-values are proportion methylated and M-values are their logit.
}
