% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\title{Association test between two variables}
\usage{
assoc(v1, v2, seed = 100)
}
\arguments{
\item{v1}{Variable 1.}

\item{v2}{Variable 2.}

\item{seed}{Random seed.}
}
\value{
p-value.
}
\description{
Association test for between variables.
}
