% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_pwys.R
\name{barplot_pwys}
\alias{barplot_pwys}
\title{Make barplots for the p-Values of Pathway Analysis}
\usage{
barplot_pwys(tab, prefix.v, name = NA, width = 10, height = 4,
  ntop = 20, direction = c("Up", "Down"), pwys_nm_size = 100)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the columns that are used in the Venn diagram. If \code{NULL} these are inferred from
\code{colnames(tab)} that end with \code{.p}.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{width}{Manual option for determining the output file width in inches.}

\item{height}{Manual option for determining the output file height in inches.}

\item{ntop}{Number of top pathways.}

\item{direction}{Direction of change.}

\item{pwys_nm_size}{The maximum number of characters for pathway names. Longer names will be truncated.}
}
\value{
ggplot object
}
\description{
Make barplots for the p-Values of pathway analysis using \pkg{ggplot2}.
}
