% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_str_lengths.R
\name{binary_str_lengths}
\alias{binary_str_lengths}
\title{Calculate lengths of each string of ones in a binary vector}
\usage{
binary_str_lengths(vv)
}
\arguments{
\item{vv}{Numeric vector of ones and zeroes only.}
}
\description{
Calculate lengths of each string of ones in a binary vector. If none are found, a numeric vector of length zero is returned.
}
