% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_directional_pvalues.R
\name{combine_directional_pvalues}
\alias{combine_directional_pvalues}
\title{Directionally combine p-values}
\usage{
combine_directional_pvalues(
  tab,
  prefix.v = NULL,
  p.suffix = "p",
  direction.suffix = "logFC",
  only.p = FALSE,
  prefix = NULL
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the columns that are used in the Venn diagram. If \code{NULL} these are inferred from
\code{colnames(tab)} that end with \code{.p}.}

\item{p.suffix}{The suffix of p-value columns, i.e. the character string after the period "."}

\item{direction.suffix}{The suffix of direction columns such as log fold-change, regression slope, correlation
coefficient, or the "Up/Down" values}

\item{only.p}{Logical; should only combined p-values be returned? If not, returns matrix with z-scores and FDRs also.}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}
}
\value{
Vector of p-values and/or direction and FDR .
}
\description{
Directionally combine two-sided test p-values.
}
