% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_by_grp.R
\name{cor_by_grp}
\alias{cor_by_grp}
\title{Correlation test in each group}
\usage{
cor_by_grp(
  object,
  phenotype,
  grp,
  method = "pearson",
  reorder.rows = TRUE,
  prefix = NULL,
  check.names = TRUE
)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{phenotype}{Numeric vector of sample characteristics (e.g. phenotypes or treatments). 
Should be same length as \code{ncol(object)}.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{method}{Character string indicating which correlation coefficient to be used for the test. One of 
\code{"pearson"}, \code{"kendall"}, or \code{"spearman"}, can be abbreviated, see \code{\link[stats]{cor.test}}.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{check.names}{Logical; should \code{names(phenotype)==rownames(object)} be checked?}
}
\value{
A data.frame of test statistics for each group and the combined statistics.
}
\description{
Two-sided correlation test in each group and combine the directional p-values.
}
