% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_null_distribution.R
\name{correct_null_distribution}
\alias{correct_null_distribution}
\title{Correct null distribution}
\usage{
correct_null_distribution(tab, prefix.v = NULL, p.suffix = "p",
  fdr.suffix = "FDR", direction.suffix = "logFC", verbose = TRUE)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the columns that are used in the Venn diagram. If \code{NULL} these are inferred from
\code{colnames(tab)} that end with \code{.p}.}

\item{p.suffix}{The suffix of p-value columns.}

\item{fdr.suffix}{The suffix of FDR columns.}

\item{direction.suffix}{The suffix of direction columns such as log fold-change, regression slope, correlation
coefficient, or the "Up/Down" values}

\item{verbose}{Print out status messages.}
}
\value{
Matrix or data.fame with corrected p-values and the corresponding FDR.
}
\description{
Correct null distribution using  \pkg{fdrtool}.
}
