% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_cor.R
\name{deseq2_cor}
\alias{deseq2_cor}
\title{Apply DESeq2's DESeq, results, and lfcShrink to one or more contrasts, and return a data.frame}
\usage{
deseq2_cor(
  dds,
  name = colnames(design(dds))[2],
  cols = c("pvalue", "padj", "log2FoldChange"),
  ncore = 1,
  shrunken = FALSE
)
}
\arguments{
\item{dds}{DESeqDataSet object.}

\item{name}{Name of the individual effect (coefficient) to test}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "SE", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "CI.L", "CI.R", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}

\item{ncore}{Number of cores to use.}

\item{shrunken}{logical, whether to shrink the slope}
}
\description{
Apply \pkg{DESeq2}'s \code{DESeq}, \code{results},and \code{lfcShrink} to one or more contrasts, and return a data.frame
}
