% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_find_all_markers.R
\name{deseq2_find_all_markers}
\alias{deseq2_find_all_markers}
\title{Find markers by DESeq2's pairwise wald tests}
\usage{
deseq2_find_all_markers(
  dds,
  grp,
  direction = c("up", "down"),
  design = NULL,
  method = c("exact", "asymptotic", "monte_carlo"),
  nsim = 1e+07 - 2,
  seed = 42,
  add.means = !is.null(grp),
  adjust.method = "BH",
  ncore = 1,
  shrunken = FALSE
)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{grp}{A named vector of group labels for each column in dds}

\item{direction}{Either "up" or "down" or both}

\item{design}{Optional design formula to set for dds}

\item{method}{Method for p-value calculation, one of "exact", "asymptotic", or "monte_carlo"}

\item{nsim}{Number of Monte Carlo simulations to perform if method is "monte_carlo"}

\item{seed}{Random seed for reproducibility}

\item{add.means}{Whether to add group-wise mean expression values to the output}

\item{adjust.method}{Method for p-value adjustment, passed to p.adjust}

\item{ncore}{Number of cores to use}

\item{shrunken}{Whether to use shrunken log2 fold changes for ranking}
}
\value{
Data frame.
}
\description{
Perform DESeq2's pairwise wald tests, find specifically up or down-regulated markers for each group.
}
