% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_lrt.R
\name{deseq2_lrt}
\alias{deseq2_lrt}
\title{Apply DESeq2's DESeq, results, and lfcShrink to one or more contrasts, and return a data.frame}
\usage{
deseq2_lrt(
  dds,
  grp = NULL,
  reduced = design(dds)[, 1],
  prefix = "Group",
  add.means = !is.null(grp),
  cols = c("stat", "pvalue", "padj"),
  ncore = 1
)
}
\arguments{
\item{dds}{DESeqDataSet object.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{reduced}{model matrix of the reduced model}

\item{prefix}{the prefix for the test result columns}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "SE", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "CI.L", "CI.R", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}

\item{ncore}{Number of cores to use.}

\item{shrunken}{logical, whether to shrink the log fold-change}
}
\description{
Apply \pkg{DESeq2}'s \code{DESeq}, \code{results},and \code{lfcShrink} to one or more contrasts, and return a data.frame
}
\details{
\code{grp} isn't needed if \code{add.means} is \code{FALSE}.
}
