% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_cor.R
\name{dream_cor}
\alias{dream_cor}
\title{Apply variancePartition's dream to one or more contrasts, perform moderated t-test, and return a table using limma's topTable}
\usage{
dream_cor(
  object,
  formula,
  pheno,
  weights = NA,
  coef = "",
  moderated = TRUE,
  cols = c("P.Value", "adj.P.Val", "logFC"),
  ncores = 1
)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have row names that are non-duplicated and non-empty.}

\item{formula}{specifies variables for the linear (mixed) model.  Must only specify covariates, since the rows of exprObj are automatically used as a response. e.g.: \code{~ a + b + (1|c)}  Formulas with only fixed effects also work, and \code{lmFit()} followed by \code{contrasts.fit()} are run.}

\item{pheno}{data.frame with columns corresponding to formula}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the 
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene 
weights with length equal to \code{nrow(object)}. Set to \code{NULL} to ignore \code{object$weights}. 
\code{weights=NA} (with length one) doesn't pass weights to \code{limma}.}

\item{coef}{column name of the mixed model}

\item{moderated}{Logical; should variancePartition::eBayes be used?}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}

\item{ncores}{number of cores}
}
\value{
Data frame.
}
\description{
Apply \pkg{variancePartition}'s \code{dream}, to one or more contrasts,  perform moderated t-test, and return
a table using \pkg{limma}'s \code{topTable}.
}
