% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_ftest_contrasts.R
\name{dream_ftest_contrasts}
\alias{dream_ftest_contrasts}
\title{Apply variancePartition's dream to one or more contrasts, perform moderated F-test, and return a table using limma's topTable}
\usage{
dream_ftest_contrasts(
  object,
  formula,
  pheno,
  contrast.v,
  weights = NA,
  grp = NULL,
  add.means = !is.null(grp),
  prefix = ""
)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{formula}{specifies variables for the linear (mixed) model.  Must only specify covariates, since the rows of exprObj are automatically used a a response. e.g.: \code{~ a + b + (1|c)}  Formulas with only fixed effects also work, and lmFit() followed by contrasts.fit() are run.}

\item{pheno}{data.frame with columns corresponding to formula}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the 
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene 
weights with length equal to\code{nrow(object)}. Set to \code{NULL} to ignore \code{object$weights}. 
\code{weights=NA} (with length one) doesn't pass weights to \code{limma}.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}
}
\value{
Data frame.
}
\description{
Apply \pkg{variancePartition}'s \code{dream}, to one or more contrasts,  perform perform moderated F-test, and return
a table using \pkg{limma}'s \code{topTable}.
}
