% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_contrasts.R
\name{edger_contrasts}
\alias{edger_contrasts}
\title{Apply edgeR's glmQLFit or glmFit, glmQLFTest, glmLRT, or glmWeightedF to one or more contrasts, and return a list}
\usage{
edger_contrasts(
  dge,
  grp = NULL,
  contrast.v,
  add.means = !is.null(grp),
  design = NULL,
  cols = c("PValue", "FDR", "logFC"),
  test = "QLFT",
  plot = FALSE
)
}
\arguments{
\item{dge}{DGEList object.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{design}{Design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be 
estimated.}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}

\item{test}{QLFT, LRT or WeightedFT.}

\item{plot}{Logical indicate if to plot the QC metrics.}
}
\description{
Apply \pkg{edgeR}'s \code{glmQLFit}, or \code{glmFit},  \code{glmQLFTest}, \code{glmLRT}, or \code{glmWeightedF} to one or more contrasts, and return
a list
}
\details{
If \code{design} is \code{NULL} and \code{grp} is given, design will be calculated as
\code{model.matrix(~0+grp)}. However, \code{grp} isn't needed if \code{design} is provided & \code{add.means}
is \code{FALSE}.
}
