% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezbacon.R
\name{ezbacon}
\alias{ezbacon}
\title{Apply \code{bacon} to z-scores}
\usage{
ezbacon(tab, z.cols = "z")
}
\arguments{
\item{tab}{Matrix-like object with statistical columns, some containing z-scores.}

\item{z.cols}{Indices or \code{\link{regexp}} with column names or column names suffix of numeric z-score columns.}
}
\value{
Data frame with bacon columns (z, p, FDR) inserted.
}
\description{
Apply \code{bacon} function from package \code{bacon} to z-scores to obtain new z-scores, p-values, & FDR.
}
