% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezchip2chip.R
\name{ezchip2chip}
\alias{ezchip2chip}
\title{Map gene IDs in geneset list to rownames of annotation data frame}
\usage{
ezchip2chip(G, annot, geneid.col = "Gene.Symbol", split.str = " /// ")
}
\arguments{
\item{G}{Geneset list such as those read in with \code{ezlimma::read.gmt} with list elements whose 3rd element
are the gene IDs.}

\item{annot}{Annotation matrix or data frame with row names and column names that has a column with gene IDs t
hat match those in \code{G} and whose row names represent the new gene IDs we want.}

\item{geneid.col}{Column name or column index of \code{annot} with gene IDs that match those in \code{G}.}

\item{split.str}{Character vector (or object which can be coerced to such) containing regular expression
to use for splitting \code{annot[, geneid.col]} when it contains multiple separated gene IDs.}
}
\value{
\code{G} with new gene IDs matching \code{rownames(annot)}.
}
\description{
Map gene IDs in geneset list to rownames of annotation data frame. The name is based on the Gene Set Enrichment
Analysis (GSEA) utility \href{https://www.gsea-msigdb.org/gsea/doc/GSEAUserGuideFrame.html?Chip2Chip_Page}{chip2chip}.
}
\details{
Both \code{rownames(annot)} and gene IDs in \code{G} are converted to upper case to aid matching.
}
