% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezgst.R
\name{ezgst}
\alias{ezgst}
\title{A wrapper function for \code{geneSetTest} with output to Excel}
\usage{
ezgst(
  stats.tab,
  G,
  feat.tab,
  name = NA,
  adjust.method = "BH",
  alternative = "mixed",
  type = "auto",
  ranks.only = TRUE,
  nsim = 9999,
  min.nfeats = 3,
  max.nfeats = 1000
)
}
\arguments{
\item{stats.tab}{A matrix-like data object with gene row names & named columns of numeric gene-wise statistics
(e.g. z-scores, t-statistics, F-statistics) by which genes can be ranked.
The row names should be the same as the row names of \code{feat.tab}.
All values must be \code{\link[base:is.finite]{finite}}.}

\item{G}{Gene set list as returned from \code{\link[ezlimma]{read_gmt}}.}

\item{feat.tab}{Table of feature (e.g. gene) statistics that the Excel table can link to.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Used for \code{mroast} and only for \code{fry} when 
\code{alternative} is not \code{"two.sided"}.}

\item{alternative}{Alternative hypothesis; must be one of\code{"two.sided"}; \code{"greater"} or \code{"less"},
or their synonyms  \code{"Up"} or \code{"Down"}.}

\item{type}{character string specifying whether the statistics are signed (t-like, \code{"t"}) or unsigned (F-like, \code{"f"}) or whether the function should make an educated guess (\code{"auto"}).
  If the statistic is unsigned, then it assume that larger statistics are more significant.}

\item{ranks.only}{logical, if \code{TRUE} only the ranks of the \code{statistics} are used.}

\item{nsim}{number of random samples to take in computing the p-value.
  Not used if \code{ranks.only=TRUE}.}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}

\item{max.nfeats}{Maximum number of features (e.g. genes) needed in a gene set for testing.}
}
\value{
Data frame of gene set statistics.
}
\description{
Test whether a set of genes is highly ranked relative to other genes in terms of a given statistic. Genes are assumed to be independent.
To find pathways whose genes have large magnitude changes, independent of direction of their change,
provide \code{abs} of gene stats, and set \code{alternative="Up"}.
It returns a data frame with statistics per gene set, and writes this to an Excel file.
The Excel file links to CSV files, which contain statistics per gene set.
}
\details{
Pathway (i.e. gene set) names are altered to be valid filenames in Windows and Linux. Numeric columns are
rounded to 3 significant figures.
}
