% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezumap.R
\name{ezumap}
\alias{ezumap}
\title{UMAP plot of first two dimensions}
\usage{
ezumap(
  object,
  pheno.df,
  name = "umap",
  pca = TRUE,
  initial_dims = nrow(pheno.df),
  config = umap::umap.defaults,
  method = c("naive", "umap-learn"),
  preserve.seed = TRUE,
  alpha = 1,
  all.size = NULL,
  facet = NULL,
  title = NULL,
  subtitle = NULL,
  rm.leg.title = FALSE,
  labels = FALSE,
  manual.color = NULL,
  manual.shape = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{Matrix-like object with features (e.g. genes) as rows and samples as columns.}

\item{pheno.df}{Data frame with rows as samples and columns as phenotypes.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{pca}{logical; Whether an initial PCA step should be performed (default: TRUE)}

\item{initial_dims}{integer; the number of dimensions that should be retained in the initial PCA step (default: 50)}

\item{config}{object of class umap.config}

\item{method}{character, implementation. Available methods are 'naive'
(an implementation written in pure R) and 'umap-learn' (requires python
package 'umap-learn')}

\item{preserve.seed}{logical, leave TRUE to insulate external code from
randomness within the umap algorithms; set FALSE to allow randomness used
in umap algorithms to alter the external random-number generator}

\item{alpha}{Transparency, passed to \code{\link[ggplot2]{geom_point}}.}

\item{all.size}{Passed to \code{\link[ggplot2]{geom_point}} \code{size} parameter to give size for all points without
appearing in legend. \code{ggplot2} default is size=2.}

\item{facet}{A formula with columns in \code{pheno.df} to facet by.}

\item{title}{Title text; suppressed if it is \code{NULL}.}

\item{subtitle}{Subtitle text; suppressed if it is \code{NULL} or \code{title} is \code{NULL}. If you'd like a
\code{subtitle} but no \code{title}, set \code{title = ""}.}

\item{rm.leg.title}{Logical indicating if legend title should be removed.}

\item{labels}{Logical, should sample labels be added next to points?}

\item{manual.color}{Vector passed to \code{\link[ggplot2:scale_manual]{scale_colour_manual}} for creating a
discrete color scale. Vector length should be equal to number of levels in mapped variable.}

\item{manual.shape}{Vector passed to \code{\link[ggplot2:scale_manual]{scale_shape_manual}} for creating a
discrete color scale. Vector length should be equal to number of levels in mapped variable.}

\item{plot}{Logical; should plot be generated?}

\item{...}{list of settings with values overwrite defaults from UMAP \code{config} or passed to \code{\link[ggplot2:aes_]{aes_string}}.}
}
\value{
Invisibly, a \code{ggplot} object. Its \code{data} element contains the first two principal components
appended to \code{pheno.df}.
}
\description{
UMAP plot of first two dimensions using \pkg{ggplot2}.
}
\details{
\code{object} must have colnames, and if \code{pheno.df}
is given, it is checked that \code{colnames(object)==rownames(pheno.df)}.
}
