% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezwsva.R
\name{ezwsva}
\alias{ezwsva}
\title{Weighted Surrogate Variable Analysis}
\usage{
ezwsva(
  y,
  design,
  n.sv = 1L,
  weight.by.sd = FALSE,
  plot = FALSE,
  prop.cutoff = 0.1,
  ...
)
}
\arguments{
\item{y}{numeric matrix giving log-expression or log-ratio values for a series of microarrays, or any object that can coerced to a matrix including \code{ExpressionSet}, \code{MAList}, \code{EList} or \code{PLMSet} objects.
        Rows correspond to genes and columns to samples.}

\item{design}{design matrix}

\item{n.sv}{number of surrogate variables required.}

\item{weight.by.sd}{logical, should the surrogate variables be especially tuned to the more variable genes?}

\item{plot}{logical. If \code{TRUE}, plots the proportion of variance explained by each surrogate variable.}

\item{prop.cutoff}{An proportion cutoff to select surrogate variables.}

\item{...}{other arguments can be included that would be suitable for \code{lmFit}.}
}
\value{
Data frame with statistics from rank products test.
}
\description{
Weighted surrogate variable analysis, with selection of number of surrogate variables.
}
