% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_screendmt.R
\name{fast_screendmt}
\alias{fast_screendmt}
\title{Post screenDMT}
\usage{
fast_screendmt(
  tab,
  cols = 1:4,
  direction = c("both", "same", "opposite"),
  reorder.rows = FALSE,
  p.adj.rate = c("FDR", "FWER"),
  prefix = NULL,
  keep.input = FALSE
)
}
\arguments{
\item{tab}{Matrix-like object with statistical and p-value columns. Only the signs of the statistics columns are used.
\code{tab} should have non-duplicated row names and should not have missing values.}

\item{cols}{Vector of column indices or names in the order of \code{c(stat1, p1, stat2, p2)}.}

\item{direction}{Character, either "same", "opposite", or default "both"}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{p.adj.rate}{Either "FDR" for false discovery rate or "FWER" for family-wise error rate, the rate controlled by the Bonferroni procedure.}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{keep.input}{Logical, should returned data frame include \code{tab[, cols]}? If so, \code{tab} must have column names and these
cannot include "chisq", "p", and "FDR" or "FWER" (depending on \code{p.adj.rate})}
}
\value{
Data frame whose rows correspond to the rows of \code{tab} with the same row names and whose columns include
\describe{
\item{direction}{diretion of change, up, down, or opposite}
\item{chisq}{Chi-square on 1 degree of freedom}
\item{p}{P-value}
\item{FDR or FWER}{Adjusted p-value}
}
}
\description{
Post screenDMT i.e. combining replication & reversed replication test results
}
\details{
Larger chi-square values are more significant. \code{tab} must have more than one row.
}
