% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_valids.R
\name{filter_valids}
\alias{filter_valids}
\title{Filter features}
\usage{
filter_valids(
  object,
  grp,
  min.prop = 2/3,
  at.least.one = FALSE,
  only.keep.vector = FALSE
)
}
\arguments{
\item{object}{A dat.frame or matrix containing data for filtering.}

\item{grp}{A character vector dictating the grouping.}

\item{min.prop}{The propotion of valid values for each group for retention.}

\item{at.least.one}{TRUE means to keep the row if min.prop is met for at least one group,
FALSE means min.prop must be met across all groups for retention.}

\item{only.keep.vector}{TRUE for returning only the logical vector indicates the rows to keep,
FALSE for returning the matix after filtering.}
}
\value{
Logical vector or data matirx.
}
\description{
Filter features that have more than a proportion missing values in one or all groups.
}
