% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pc_curve_elbow.R
\name{find_pc_curve_elbow}
\alias{find_pc_curve_elbow}
\title{Identify and plot the elbow point of the PC curve}
\usage{
find_pc_curve_elbow(
  object,
  method = 1,
  maxPC = 50,
  plot = TRUE,
  name = NA,
  title = "Importance of PCs",
  width = 6,
  height = 4
)
}
\arguments{
\item{method}{1 or 2. 1 = use \code{\link[pathviewr]{find_curve_elbow}}, 2 = use \code{\link[PCAtools]{findElbowPoint}}, default is 1.}

\item{maxPC}{maximum PCs to be considered. Default is 50.}

\item{title}{figure title.}

\item{width}{figure width.}

\item{height}{figure height.}
}
\value{
the PC number at elbow point.
}
\description{
Identify and plot the elbow point of the PC curve
}
