% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_cluster_markers.R
\name{find_top_cluster_markers}
\alias{find_top_cluster_markers}
\title{Find top cluster markers}
\usage{
find_top_cluster_markers(
  cluster.markers,
  ntop = 5,
  padj.cutoff = 0.05,
  pct.cutoff = 0.25
)
}
\arguments{
\item{cluster.markers}{Data.frame generrated bt Seurat FindAllMarkers}

\item{ntop}{Number of top markers for each cluster}

\item{padj.cutoff}{Cutoff of adjusted p-values}

\item{pct.cutoff}{Cutoff of percentage of expressing cells}
}
\value{
Data.frame
}
\description{
Find top cluster markers.
}
