% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_in_grp.R
\name{hclust_in_grp}
\alias{hclust_in_grp}
\title{Hierarchical Clustering in each group}
\usage{
hclust_in_grp(
  object,
  grp,
  sc = "ctr",
  clip = NA,
  dist.method = "euclidean",
  hc.method = "ward.D2"
)
}
\arguments{
\item{object}{Matrix-like object with features (e.g. genes) as rows and samples as columns.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{sc}{Row scaling. Should rows be centered ('ctr'), z-scored ('z'), or neither ('none').}

\item{clip}{Values with magnitude > \code{clip} are reset to value \code{clip}. If given, must be > 0.}

\item{dist.method}{The distance measure to be used.}

\item{hc.method}{The agglomeration method to be used.}
}
\value{
Vector of the labels.
}
\description{
Hierarchical cluster analysis in each group.
}
