% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_normal.R
\name{impute_normal}
\alias{impute_normal}
\title{Impute missing values by random numbers drawn from a normal distribution}
\usage{
impute_normal(object, width = 0.3, downshift = 1.8, seed = 100)
}
\arguments{
\item{object}{Data frame or matrix containing filtered and log-transformed data.}

\item{width}{Scale factor for the standard deviation of imputed distribution relative to the sample standard deviation.}

\item{downshift}{Down-shifted the mean of imputed distribution from the sample mean, in units of sample standard deviation.}

\item{seed}{Random seed}
}
\value{
A matix with imupted values
}
\description{
Impute missing values by random numbers drawn from a normal distribution which has a mean down-shifted and standard deviation shrinked from the sample distribution.
}
