% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ftest_contrasts.R
\name{limma_ftest_contrasts}
\alias{limma_ftest_contrasts}
\title{Apply limma's lmFit, contrasts.fit, & eBayes to one or more contrasts, perform moderated F-test, and return a table}
\usage{
limma_ftest_contrasts(object, grp = NULL, contrast.v,
  add.means = !is.null(grp), weights = NA, design = NULL,
  prefix = "", trend = FALSE, block = NULL, correlation = NULL)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the 
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene 
weights with length equal to\code{nrow(object)}. Set to \code{NULL} to ignore \code{object$weights}. 
\code{weights=NA} (with length one) doesn't pass weights to \code{limma}.}

\item{design}{Design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be 
estimated.}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{trend}{Logical; should an intensity-trend be allowed for the prior variance? Default is that the prior variance 
is constant.}

\item{block}{Vector specifying a blocking variable on the samples. Has length = \code{ncol(object)}.}

\item{correlation}{Inter-duplicate or inter-technical replicate correlation.}
}
\value{
Data frame.
}
\description{
Apply \pkg{limma}'s \code{lmFit}, \code{contrasts.fit}, & \code{eBayes} to one or more contrasts,  perform moderated F-test, and return
a table.
}
\details{
If \code{design} is \code{NULL} and \code{grp} is given, design will be calculated as
\code{model.matrix(~0+grp)}. However, \code{grp} isn't needed if \code{design} is provided & \code{add.means}
is \code{FALSE}.
}
\references{
McCarthy DJ & Smyth GK (2009). Testing significance relative to a fold-change threshold is a TREAT.
Bioinformatics 25, 765-771.
}
\seealso{
\code{\link[limma]{lmFit}}; \code{\link[limma]{eBayes}}.
}
