% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_treat_find_all_markers.R
\name{limma_treat_find_all_markers}
\alias{limma_treat_find_all_markers}
\title{Find markers for each group by limma's pairwise moderated t-tests over a logFC cutoff}
\usage{
limma_treat_find_all_markers(
  object,
  grp,
  direction = c("up", "down"),
  treat.lfc = log2(1.2),
  design = NULL,
  add.means = !is.null(grp),
  adjust.method = "BH",
  weights = NA,
  trend = FALSE,
  block = NULL,
  correlation = NULL,
  moderated = TRUE
)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have row names that are non-duplicated and non-empty.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{direction}{Either "up" or "down" or both}

\item{treat.lfc}{Vector of logFC passed to \code{\link[limma:lmFit]{treat}} \code{lfc}. It is recycled as needed to match
rows of \code{object}. If given, \code{length(contrast.v)} must be 1. McCarthy & Smyth suggest a 10 percent fold-change,
which is \code{treat.lfc=log2(1.1)}. Limma Treat tests for logFC's outside the interval [-treat.lfc, treat.lfc].}

\item{design}{Design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be 
estimated.}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the 
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene 
weights with length equal to \code{nrow(object)}. Set to \code{NULL} to ignore \code{object$weights}. 
\code{weights=NA} (with length one) doesn't pass weights to \code{limma}.}

\item{trend}{Logical; should an intensity-trend be allowed for the prior variance? Default is that the prior variance 
is constant.}

\item{block}{Vector specifying a blocking variable on the samples. Has length = \code{ncol(object)}. 
Must be \code{NULL} if \code{ndups > 1}.}

\item{correlation}{Inter-duplicate or inter-technical replicate correlation. Must be given if 
\code{ndups>1} or \code{!is.null(block)}.}

\item{moderated}{Logical; should \code{\link[limma]{eBayes}} be used? Otherwise an unmoderated version for 
\pkg{limma} to produce ordinary least squares statistics is used.}
}
\value{
Data frame.
}
\description{
Perform limma's pairwise moderated t-tests, find specifically up or down-regulated markers for each group over a logFC cutoff.
}
