% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_trend_plot.R
\name{limma_trend_plot}
\alias{limma_trend_plot}
\title{Mean-variance trend plot}
\usage{
limma_trend_plot(
  object,
  design = NULL,
  name = NA,
  span = 0.5,
  alpha = 0.5,
  width = 7,
  height = 7,
  ...
)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a series of arrays, with rows corresponding to genes and columns to samples.
  Any type of data object that can be processed by \code{\link[limma]{getEAWP}} is acceptable.}

\item{design}{the design matrix of the microarray experiment, with rows corresponding to samples and columns to coefficients to be estimated.
  Defaults to \code{object$design} if that is non-NULL, otherwise to the unit vector meaning that all samples will be treated as replicates of a single treatment group.}

\item{name}{file name for plot}

\item{span}{
    width of the smoothing window used for the lowess mean-variance trend.
    Expressed as a proportion between 0 and 1.}

\item{alpha}{Transparency, passed to \code{\link[ggplot2]{geom_point}}.}

\item{width, height}{the width and height of the graphics region in
    inches.  The default values are \code{7}.}

\item{...}{Passed to \code{\link[limma]{lmFit}}.}
}
\description{
Estimate the mean-variance relationship and plot its trend using \code{limma} and \code{ggplot2}.
}
