% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file_links.R
\name{make_file_links}
\alias{make_file_links}
\title{Make file links for R Markdown}
\usage{
make_file_links(
  path = ".",
  pattern = NULL,
  recursive = FALSE,
  full.names = TRUE
)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link[base]{getwd}()}.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is performed.  Missing
    values will be ignored.  Elements with a marked encoding will
    be converted to the native encoding (and if that fails, considered
    non-existent).}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}
}
\value{
Vector of file links.
}
\description{
Make file links for R Markdown.
}
