% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_piechart.R
\name{make_piechart}
\alias{make_piechart}
\title{Make a pie chart}
\usage{
make_piechart(tab, Count = "Count", Category = "Category", title = NULL)
}
\arguments{
\item{tab}{a table of counts and category}

\item{Count}{column name of counts in tab}

\item{Category}{column name of categories in tab}

\item{title}{title of the chart}
}
\value{
ggplot2 object.
}
\description{
Make a pie chart with labels of counts
}
