% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smpdb_glist.R
\name{make_smpdb_glist}
\alias{make_smpdb_glist}
\title{Make a feature list for SMPDB pathways.}
\usage{
make_smpdb_glist(
  smpdb.tab,
  annot,
  id.cols.smpdb = c("Metabolite.ID", "ChEBI.ID", "KEGG.ID"),
  id.cols.annot = c("HMDB", "ChEBI", "KEGG"),
  pwys.name.col = "Pathway.Name",
  pwys.id.col = "SMPDB.ID",
  pwys.type.col = "Pathway.Type",
  pwys.type = "Metabolic"
)
}
\arguments{
\item{smpdb.tab}{A dataframe of SMPDB data, whith column names in \code{id.cols.smpdb}, \code{pwys.name.col}, and \code{pwys.id.col}.}

\item{annot}{Metabolite annotation, with column names in \code{id.cols.annot}}

\item{id.cols.smpdb}{Column names in \code{smpdb.tab} for HMDB, ChEBI, and KEGG IDs.}

\item{id.cols.annot}{Column names in \code{annot} for HMDB, ChEBI, and KEGG IDs.}

\item{pwys.name.col}{Column name in \code{smpdb.tab} for pathway names.}

\item{pwys.id.col}{Column name in \code{smpdb.tab} for pathway IDs.}

\item{pwys.type.col}{Column name in \code{smpdb.tab} for pathway types.}

\item{pwys.type}{Type of pathways to use. Default is "Metabolic".}
}
\value{
A list
}
\description{
Make a feature list for SMPDB pathways.
}
