% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_by_adist.R
\name{map_by_adist}
\alias{map_by_adist}
\title{Map two character vectors by approximate string distances}
\usage{
map_by_adist(x, y, ignore.case = TRUE, max.norm.dist = 0.2)
}
\arguments{
\item{x}{a character vector.}

\item{y}{a character vector.}

\item{ignore.case}{a logical. If TRUE, case is ignored for computing the distances.}

\item{max.norm.dist}{maximal normalized distance between 0 and 1.}
}
\value{
Two-column Data frame of which each row is the mapped elements in x and y.
}
\description{
Map two character vectors by generalized Levenshtein (edit) distance i.e. approximate string distances.
}
