% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_t2g.R
\name{map_t2g}
\alias{map_t2g}
\title{Map pathway term to gene IDs to another ID type}
\usage{
map_t2g(
  t2g,
  annot,
  sep.str = " /// ",
  symbol.col = "Gene.Symbol",
  fixed = FALSE
)
}
\arguments{
\item{t2g}{TERM TO GENE mapping, a data.frame of 2 column with term and gene}

\item{annot}{annotation for the features that has a column of the same type as in gene set list \emph{G}.}

\item{sep.str}{strings that separates symbols if there are multiple symbols for a feature.}

\item{symbol.col}{column name or index for the symbol column in \emph{annot}.}

\item{fixed}{logical passed to \code{strsplit}; if \code{TRUE}, match \code{sep.str} exactly, otherwise use regular expression.}
}
\value{
TERM TO GENE data.frame with the symbols replaced by the mapped IDs.
}
\description{
Map pathway term to gene with old gene IDs (e.g. entrez IDs) to another ID type (e.g. gene symbols) using a table with both
annotation types. The original IDs can be separated within a column of annot with a string, e.g. " /// ".
The other IDs should be the rownames of annot. The idea for this is based on the Gene Set Enrichment
Analysis (GSEA) utility \href{https://www.gsea-msigdb.org/gsea/doc/GSEAUserGuideFrame.html?Chip2Chip_Page}{chip2chip}.
}
\details{
Gene annotations are transformed to upper-case to avoid missing matches of different cases.
}
