% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_met_nms.R
\name{match_met_nms}
\alias{match_met_nms}
\title{Match names of chemicals}
\usage{
match_met_nms(x, table)
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\value{
Vector of the positions of (first) matches of its first argument in its second.
}
\description{
Match names of chemicals, which are often common, non-systematic names.
}
