% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_enrichFLY.R
\name{multi_enrichFLY}
\alias{multi_enrichFLY}
\title{multi_enrichFLY}
\usage{
multi_enrichFLY(
  object,
  G,
  annot,
  sep.str = " /// ",
  symbol.col = "symbol",
  grp = NULL,
  contrast.v = ncol(design),
  design = NULL,
  weights = NULL,
  trend = FALSE,
  block = NULL,
  correlation = NULL,
  adjust.method = c("BH", "none"),
  min.nfeats = 3,
  max.nfeats = 1000,
  pvalueCutoff = 0.25,
  qvalueCutoff = 1
)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to
features (e.g. genes) and columns to samples. Must have row names that are non-duplicated and non-empty.}

\item{G}{Gene set list as returned from \code{\link{read_gmt}}.}

\item{annot}{annotation for the features that has a column of the same type as in gene set list \code{G}.}

\item{sep.str}{strings that separates symbols if there are multiple symbols for a feature.}

\item{symbol.col}{column name or index for the symbol column in \code{annot}.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as
\code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{design}{Design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be
estimated.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene
weights with length equal to \code{nrow(object)}.}

\item{trend}{Logical; should an intensity-trend be allowed for the prior variance? Default is that the prior variance
is constant.}

\item{block}{Vector specifying a blocking variable on the samples. Has length = \code{ncol(object)}.
Must be \code{NULL} if \code{ndups > 1}.}

\item{correlation}{Inter-duplicate or inter-technical replicate correlation. Must be given if
\code{ndups>1} or \code{!is.null(block)}.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options,}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}

\item{max.nfeats}{Maximum number of features (e.g. genes) needed in a gene set for testing.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{qvalueCutoff}{Cutoff value of qvalue.}
}
\value{
A list of 3 \code{compareClusterResult} instance for up-regulated, down-regulated, and mixed-regulated enriched gene sets.
}
\description{
Enrichment analyzer using \code{\link[limma]{roast}} with function \code{fry}
}
