% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_pval2zlist.R
\name{multi_pval2zlist}
\alias{multi_pval2zlist}
\title{Convert multiple columns of p-values to list of named and sorted Z-scores}
\usage{
multi_pval2zlist(
  tab,
  prefix.v = NULL,
  p.suffix = "p",
  direction.suffix = "logFC",
  sort = TRUE,
  decreasing = TRUE,
  nm = rownames(tab)
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the columns that are used in the Venn diagram. If \code{NULL} these are inferred from
\code{colnames(tab)} that end with \code{.p}.}

\item{p.suffix}{The suffix of p-value columns, i.e. the character string after the period "."}

\item{direction.suffix}{The suffix of direction columns such as log fold-change, regression slope, correlation
coefficient, or the "Up/Down" values}

\item{sort}{Logical. Should the Z-scores be sorted?}

\item{decreasing}{Logical. Should the sort be increasing or decreasing?}

\item{nm}{Names of the Zscores.}
}
\value{
List of Z-scores.
}
\description{
Convert multiple columns (from a table-like object) of p-values from two-sided tests to list of named and sorted Z-scores.
}
