% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggsave.R
\name{my_ggsave}
\alias{my_ggsave}
\title{Wrapper for ggplot2::ggsave}
\usage{
my_ggsave(
  name,
  plot = get_last_plot(),
  width = NA,
  height = NA,
  units = "in",
  dpi = 300,
  limitsize = FALSE,
  bg = "white",
  ...
)
}
\arguments{
\item{name}{The name of plots (without the format extension)}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{width, height}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}

\item{units}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Only applies when converting pixel units,
as is typical for raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\value{
A character vector of file names of the PNG and PDF
}
\description{
Call ggsave twice to save the plot as both PNG and PDF
}
