% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cutadapt_logs.R
\name{parse_cutadapt_logs}
\alias{parse_cutadapt_logs}
\title{Parse cutadapt log files}
\usage{
parse_cutadapt_logs(name = "cutadapt_summary", path = ".")
}
\arguments{
\item{name}{the file name of result.}

\item{path}{the full path name for the log files.}
}
\value{
Data frame with summary of cutadapt.
}
\description{
Parse cutadapt log files.
}
