% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_featureCounts_logs.R
\name{parse_featureCounts_logs}
\alias{parse_featureCounts_logs}
\title{Parse featureCounts log file}
\usage{
parse_featureCounts_logs(file, name = "featureCounts_summary")
}
\arguments{
\item{file}{the log file name including the full path.}

\item{name}{the file name of result.}
}
\value{
Data frame with summary of featureCounts.
}
\description{
Parse featureCounts log file.
}
