% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_kallisto_logs.R
\name{parse_kallisto_logs}
\alias{parse_kallisto_logs}
\title{Parse kallisto log files}
\usage{
parse_kallisto_logs(name = "kallisto_summary", path = ".")
}
\arguments{
\item{name}{the file name of result.}

\item{path}{the full path name for the kallisto log files.}
}
\value{
Data frame with summary of kallisto alignment.
}
\description{
Parse kallisto log files.
}
