% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_star_logs.R
\name{parse_star_logs}
\alias{parse_star_logs}
\title{Parse star log files}
\usage{
parse_star_logs(name = "star_summary", path = ".")
}
\arguments{
\item{name}{the file name of result.}

\item{path}{the full path name for the log files.}
}
\value{
Data frame with summary of star alignment.
}
\description{
Parse star log files.
}
