% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot logit missing probability by average log abundance}
\usage{
plot_missing(mat, name = NA, main = "", alpha = 1)
}
\arguments{
\item{mat}{Numeric matrix of log abundance.}

\item{name}{file names of the plot.}

\item{main}{title of the plot.}

\item{alpha}{Transparency of color.}
}
\value{
NULL.
}
\description{
Plot logit missing probability by average log abundance.
}
