% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankprod.R
\name{rankprod}
\alias{rankprod}
\title{Rank products of a matrix where larger statistics have stronger rank}
\usage{
rankprod(mat, nsim = 1e+07 - 1, reorder.rows = TRUE, prefix = NULL)
}
\arguments{
\item{mat}{Numerix matrix with columns holding statistics per comparison & rows are analytes.}

\item{nsim}{Numeric number of simulations.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}
}
\value{
Data frame with statistics from rank products test.
}
\description{
Rank products of a matrix. Larger statistics have stronger rank.
}
