% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcomp.R
\name{revcomp}
\alias{revcomp}
\title{converts a DNA sequence into its reverse-complement counterpart}
\usage{
revcomp(x)
}
\arguments{
\item{x}{a character vector of strings of DNA sequences}
}
\value{
a character vector of strings of the reverse-complement DNA sequences
}
\description{
converts a DNA sequence into its reverse-complement counterpart.
}
