% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bsgenome_regions.R
\name{sample_bsgenome_regions}
\alias{sample_bsgenome_regions}
\title{Rand seq}
\usage{
sample_bsgenome_regions(BSgenome, nsamp, flank, seed = 123)
}
\arguments{
\item{BSgenome}{BSgenome object.}

\item{nsamp}{Numeric number of samples.}

\item{flank}{Numeric length of flank on each side of region.}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}
}
\description{
Rand seq.
}
