% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_top.R, R/select_top_2sided.R
\name{select_top}
\alias{select_top}
\title{Select equal numbers of top features from the topTable}
\usage{
select_top(
  tab,
  prefix.v = NULL,
  p.suffix = "p",
  direction.suffix = "logFC",
  ntop = 50,
  ntop.offset = 5,
  each = TRUE
)

select_top(
  tab,
  prefix.v = NULL,
  p.suffix = "p",
  direction.suffix = "logFC",
  ntop = 50,
  ntop.offset = 5,
  each = TRUE
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the columns that are used in the Venn diagram. If \code{NULL} these are inferred from
\code{colnames(tab)} that end with \code{.p}.}

\item{p.suffix}{The suffix of p-value columns, i.e. the character string after the period "."}

\item{direction.suffix}{The suffix of direction columns such as log fold-change, regression slope, correlation
coefficient, or the "Up/Down" values}

\item{ntop}{Number of top features}

\item{ntop.offset}{Offset for ntop}

\item{each}{Logical TRUE for select equal number of top features from each comparison}
}
\value{
Vector of the top feature's IDs (the rownames of topTable).

Vector of the top feature's IDs (the rownames of topTable).
}
\description{
Select equal numbers of top features from the topTable based on p-values.

Select equal numbers of top features from the topTable based on p-values.
}
