% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_figure.R
\name{show_figure}
\alias{show_figure}
\title{Wrapper for print out PNG}
\usage{
show_figure(path, title, caption, one_col_width = c("50\%", "80\%"))
}
\arguments{
\item{path}{The vector of PNG file names (<=3 files)}

\item{title}{The vector of figure titles}

\item{caption}{The vector of figure captions}

\item{one_col_width}{The width of one-column figure}
}
\description{
Print out PNG in in HTML format. This is useful for R Markdown reports.
}
