% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_table.R
\name{show_table}
\alias{show_table}
\title{Wrapper for knitr::kable}
\usage{
show_table(x, caption, digits = 3)
}
\arguments{
\item{x}{The table to be printed}

\item{caption}{A caption for the table}

\item{digits}{The number of digits to display}
}
\description{
Call knitr::kable to print out the table in HTML format. This is useful for R Markdown reports.
}
