% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_rankprod.R
\name{signed_rankprod}
\alias{signed_rankprod}
\title{Signed rank products of a two column matrix where larger statistics have stronger rank in the same or opposite direction}
\usage{
signed_rankprod(
  mat,
  nsim = 1e+07 - 1,
  same.dirction = FALSE,
  reorder.rows = TRUE,
  prefix = NULL,
  seed = 100
)
}
\arguments{
\item{mat}{Numeric matrix with two columns holding statistics per comparison & rows are analytes.}

\item{nsim}{Numeric number of simulations.}

\item{same.dirction}{Logical indicates whether the two ranks should be in the same direction.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{seed}{Random seed.}
}
\value{
Data frame with statistics from signed rank products test.
}
\description{
Signed rank products of a two column matrix. Larger statistics have stronger rank in the same or opposite direction.
}
