% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_ranksum.R
\name{signed_ranksum}
\alias{signed_ranksum}
\title{Signed rank sum of a matrix where larger statistics have stronger rank in the same direction}
\usage{
signed_ranksum(mat, nsim = 1e+07 - 1, reorder.rows = TRUE,
  prefix = NULL, seed = 100)
}
\arguments{
\item{mat}{Numeric matrix with columns holding statistics per comparison & rows are analytes.}

\item{nsim}{Numeric number of simulations.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{seed}{Random seed.}
}
\value{
Data frame with statistics from signed rank sum test.
}
\description{
Signed rank sum of a matrix. Larger statistics have stronger rank in the same direction.
}
