% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_hist_qq.R
\name{signif_hist_qq}
\alias{signif_hist_qq}
\title{Plot histograms of significance (p-value & FDR) columns}
\usage{
signif_hist_qq(
  tab,
  p.suffix = "p",
  fdr.suffix = "FDR",
  sep = ".",
  pi0 = FALSE,
  method = c("lfdr", "convest")
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{p.suffix}{Suffix for p-value columns. P-value column names cannot be duplicated.}

\item{fdr.suffix}{Suffix for FDR columns. Set to \code{NA} if no FDR columns. FDR column names cannot be duplicated.}

\item{sep}{Separator for column names before \code{p} or \code{FDR}, passed to \code{\link{extract_prefix}}.
If not found, it is assumed to be \code{NA}.}

\item{pi0}{Logical indicating if proportion of null hypotheses should be calculated per p-value histogram.}

\item{method}{Method for estimating pi0, passed to \code{\link[limma:propTrueNull]{limma::propTrueNull}}. Ignored if \code{!pi0}.}
}
\value{
List of ggplots.
}
\description{
Plot histograms of significance (p-value & FDR) and QQ (p-value) columns.
}
\details{
Some p-value columns must be identifiable using \code{p.suffix}. If \code{!is.na(fdr.suffix)}, FDR
colnames must have same prefix.
}
