% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_fisher_test.R
\name{venn_fisher_test}
\alias{venn_fisher_test}
\title{Fisher exact test for 2-way Venn}
\usage{
venn_fisher_test(mat.sig)
}
\arguments{
\item{mat.sig}{Numeric matrix with {1, -1, 0} for significantly up/down, insignificant features.}
}
\value{
Data frame with statistics from Fisher exact test.
}
\description{
Fisher exact test for enrichment of overlap in 2-way Venn.
}
