% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_2prop_pooled.R
\name{z_2prop_pooled}
\alias{z_2prop_pooled}
\title{Two-proportion z-test (pooled) for non-zero null}
\usage{
z_2prop_pooled(
  x1,
  x2,
  n1,
  n2,
  delta = 0,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x1}{number of successes in group 1.}

\item{x2}{number of successes in group 2.}

\item{n1}{total number of games in group 1.}

\item{n2}{total number of games in group 2.}

\item{delta}{null hypothesis of the difference in proportion.}

\item{alternative}{a character string specifying the alternative
    hypothesis, must be one of \code{"two.sided"} (default),
    \code{"greater"} or \code{"less"}.  You can specify just the initial
    letter.  Only used for testing the null that a single proportion
    equals a given value, or that two proportions are equal; ignored
    otherwise.}
}
\value{
a number of p-value.
}
\description{
Two-proportion z-test (pooled) for non-zero null.
}
