% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_contrasts.R
\name{deseq2_contrasts}
\alias{deseq2_contrasts}
\title{Apply DESeq2's DESeq, results, and lfcShrink to one or more contrasts, and return a data.frame}
\usage{
deseq2_contrasts(
  dds,
  grp = NULL,
  contrast.v,
  add.means = !is.null(grp),
  cols = c("pvalue", "padj", "log2FoldChange"),
  shrunken = TRUE,
  add_cooks = FALSE,
  independentFiltering = FALSE,
  parallel = TRUE,
  ncore = 1,
  alpha = 0.1,
  lfc_shrink_type = c("ashr", "apeglm", "normal")
)
}
\arguments{
\item{dds}{DESeqDataSet object.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "SE", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "CI.L", "CI.R", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}

\item{shrunken}{logical, whether to shrink the log fold-change.}

\item{add_cooks}{logical, whether to include Cook's cutoff. \code{DESeq2} default is \code{TRUE}.}

\item{independentFiltering}{logical, whether independent filtering should be
applied automatically}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel
execution using \code{BiocParallel}, see next argument \code{BPPARAM}}

\item{ncore}{Number of cores to use.}

\item{alpha}{the significance cutoff used for optimizing the independent
filtering (by default 0.1). If the adjusted p-value cutoff (FDR) will be a
value other than 0.1, \code{alpha} should be set to that value.}

\item{lfc_shrink_type}{character string, method for \code{lfcShrink}. It is recommended to not use method \code{"normal"}.}
}
\description{
Apply \pkg{DESeq2}'s \code{DESeq}, \code{results},and \code{lfcShrink} to one or more contrasts, and return a data.frame
}
\details{
\code{grp} isn't needed if \code{add.means} is \code{FALSE}. To match \code{DESeq2} default, set \code{add_cooks = independentFiltering = TRUE}.
This function's default values are chosen to be back-compatible with previous versions.
}
