% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_find_all_markers.R
\name{deseq2_find_all_markers}
\alias{deseq2_find_all_markers}
\title{Find markers by DESeq2's pairwise wald tests}
\usage{
deseq2_find_all_markers(
  dds,
  grp,
  direction = c("up", "down"),
  design = NULL,
  add.means = !is.null(grp),
  adjust.method = "BH",
  ncore = 1,
  shrunken = TRUE
)
}
\arguments{
\item{dds}{DESeqDataSet object.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{direction}{Either "up" or "down" or both}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{ncore}{Number of cores to use.}

\item{shrunken}{logical, whether to shrink the log fold-change}
}
\value{
Data frame.
}
\description{
Perform DESeq2's pairwise wald tests, find specifically up or down-regulated markers for each group.
}
