% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Mode estimation of a continuous variable}
\usage{
find_mode(x, xlab = NULL, main = "", by = NULL, bin = 20)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{xlab}{these arguments to \code{title} have useful
    defaults here.}

\item{main}{these arguments to \code{title} have useful
    defaults here.}

\item{bin}{a single number giving the number of cells for the histogram.}
}
\value{
Data frame with the values and densities of estimated modes.
}
\description{
Kernel density estimation to find the modes of a continuous variable.
}
