% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Mode estimation of a continuous variable}
\usage{
find_mode(x, xlab = NULL, main = "", by = NULL, bin = 20, plot = TRUE)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{by}{increment of the bins i.e. the width of bins.}

\item{bin}{the number of bins for the histogram.}

\item{plot}{logical for ploting the histogram.}
}
\value{
Data frame with modes values and densities.
}
\description{
Kernel density estimation of modes of a continuous variable.
}
