% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_and_heat.R
\name{hclust_and_heat}
\alias{hclust_and_heat}
\title{Hierarchical Clustering & visulization by heatmap}
\usage{
hclust_and_heat(object, annot, sc = "z", clip = NA,
  dist.method = "euclidean", hc.method = "ward.D2", deepSplit = 1,
  minClusterSize = 10, verbose = TRUE, pheno.df = NULL,
  labrows = "", labcols = "", color.v = NULL,
  main = "Log2 Expression", name = "hclust_heat",
  reorder_cols = FALSE, gaps_col = NULL, width = NA, height = NA,
  plot = TRUE)
}
\arguments{
\item{object}{Matrix-like object with features (e.g. genes) as rows and samples as columns.}

\item{annot}{Feature annotation.}

\item{sc}{Row scaling. Should rows be centered ('ctr'), z-scored ('z'), or neither ('none').}

\item{clip}{Values with magnitude > \code{clip} are reset to value \code{clip}. If given, must be > 0.}

\item{dist.method}{The distance measure to be used.}

\item{hc.method}{The agglomeration method to be used.}

\item{deepSplit}{For method "hybrid", can be either logical or integer in the range 0 to 4. For method
"tree", must be logical. In both cases, provides a rough control over sensitivity to cluster splitting.
The higher the value (or if \code{TRUE}), the more and smaller clusters will be produced. For the
"hybrid" method, a finer control can be achieved via \code{maxCoreScatter} and \code{minGap} below.}

\item{minClusterSize}{Minimum cluster size.  }

\item{verbose}{Logical; print pruning messages to console?}

\item{pheno.df}{Data frame with rows as samples and columns as phenotypes.}

\item{labrows}{Labels for rows, e.g. gene symbols. This can be of length 1 (in which case it is recycled),
of length \code{nrow(object)}.}

\item{labcols}{Labels for columns. This can be of length 1 (in which case it is recycled),
of length \code{ncol(object)}.}

\item{color.v}{Color palette for heatmap. If \code{NULL}, it's set to
\code{colorRampPalette(rev(brewer.pal(n=9, name='RdYlBu')))(50)}.}

\item{main}{Main title of plot.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{reorder_cols}{Logical; should columns be reordered with hierarchical clustering?}

\item{gaps_col}{vector of column indices after which to insert gaps.}

\item{width}{Manual option for determining the output file width in inches.}

\item{height}{Manual option for determining the output file height in inches.}

\item{plot}{Logical; should plot be generated?}
}
\value{
A data.frame of the clusters assingment and feature annotation.
}
\description{
Hierarchical cluster analysis and visualization by \pkg{ezlimmaplot} \code{ezheat}.
}
