% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_find_all_markers.R
\name{limma_find_all_markers}
\alias{limma_find_all_markers}
\title{Find markers for each group by limma's pairwise moderated t-tests}
\usage{
limma_find_all_markers(
  object,
  grp,
  direction = c("up", "down"),
  design = NULL,
  method = c("exact", "asymptotic", "monte_carlo"),
  nsim = 1e+07 - 2,
  seed = 42,
  add.means = !is.null(grp),
  adjust.method = "BH",
  weights = NA,
  trend = FALSE,
  block = NULL,
  correlation = NULL,
  moderated = TRUE
)
}
\arguments{
\item{object}{A numeric matrix or voom object}

\item{grp}{A named vector of group labels for each column in object}

\item{direction}{Either "up" or "down" or both}

\item{design}{Optional design matrix to set for the linear model}

\item{method}{Method for p-value calculation, one of "exact", "asymptotic", or "monte_carlo"}

\item{nsim}{Number of Monte Carlo simulations to perform if method is "monte_carlo"}

\item{seed}{Random seed for reproducibility}

\item{add.means}{Whether to add group-wise mean expression values to the output}

\item{adjust.method}{Method for p-value adjustment, passed to p.adjust}

\item{weights}{Optional observation weights for limma}

\item{trend}{Whether to use mean-variance trend for limma}

\item{block}{Optional blocking factor for limma}

\item{correlation}{Optional correlation structure for limma}

\item{moderated}{Whether to use moderated t-statistics for limma}
}
\value{
Data frame.
}
\description{
Perform limma's pairwise moderated t-tests, find specifically up or down-regulated markers for each group.
}
