% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gsea_barplot.R
\name{make_gsea_barplot}
\alias{make_gsea_barplot}
\title{Make a barplot for GSEA results}
\usage{
make_gsea_barplot(gseaObj, ntop = 10)
}
\arguments{
\item{gseaObj}{GSEA object}

\item{ntop}{number of up and down top patwys to shown}
}
\value{
ggplot2 object.
}
\description{
Make a barplot showing the NES of top pathways of GSEA results
}
