% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gsea_cnet.R
\name{make_gsea_cnet}
\alias{make_gsea_cnet}
\title{Make cnets plot for a list of GSEA results}
\usage{
make_gsea_cnet(gseaRes, fc, resolution = 1, label = "Z-scores")
}
\arguments{
\item{gseaRes}{a named list of GSEA object}

\item{fc}{a list of Z-scores with the same names as gseaRes}
}
\value{
a list of ggplot2 objects.
}
\description{
Make cnet plots for the community-representing pathways
}
