% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_assoc.R
\name{multi_assoc}
\alias{multi_assoc}
\title{Association tests for two groups of variables}
\usage{
multi_assoc(
  object1,
  object2,
  seed = 100,
  plot = TRUE,
  name = "assoc_pvals",
  width = 8,
  height = 7
)
}
\arguments{
\item{object1}{Vector, matrix or data.frame of group 1 variables, each column is 1 variable.}

\item{object2}{Vector, matrix or data.frame of group 2 variables, each column is 1 variable.}

\item{seed}{Random seed.}

\item{plot}{Logical for plotting the p-values.}

\item{name}{Name of the plot file.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}
}
\value{
Matrix of p-values.
}
\description{
Association tests for two groups of variables
}
