% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_pval2z.R
\name{multi_pval2z}
\alias{multi_pval2z}
\title{Convert the table of p-values to Z-scores}
\usage{
multi_pval2z(tab, prefix.v = NULL, p.suffix = "p",
  direction.suffix = "logFC")
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the columns that are used in the Venn diagram. If \code{NULL} these are inferred from
\code{colnames(tab)} that end with \code{.p}.}

\item{p.suffix}{The suffix of p-value columns}

\item{direction.suffix}{The suffix ofdirection columns such as log fold-change, regression slope, correlation coefficient, or the"Up/Down" valuses.}
}
\value{
Matrix of Z-scores.
}
\description{
Convert the table of two-sided test p-values to Z-scores.
}
