% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_pc_regression.R
\name{multivariate_pc_regression}
\alias{multivariate_pc_regression}
\title{Multivariate regression for meta variables and principal components}
\usage{
multivariate_pc_regression(
  object,
  pheno.df,
  metavars,
  num.pc = 10,
  plot = TRUE,
  name = "assoc_with_pc_pvals",
  width = 8,
  height = 7,
  xrot = 0,
  yrot = 0
)
}
\arguments{
\item{object}{Matrix-like object with features (e.g. genes) as rows and samples as columns.}

\item{pheno.df}{Data frame with rows as samples and columns as phenotypes.}

\item{metavars}{A vector of column names in pheno.df.}

\item{num.pc}{Number of principal components}

\item{plot}{Logical for plotting the p-values.}

\item{name}{Name of the plot file.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}

\item{xrot}{rot angle (in degrees) by which the x- axis labels are to be rotated}

\item{yrot}{rot angle (in degrees) by which the y- axis labels are to be rotated}
}
\value{
Matrix of p-values.
}
\description{
Multivariate regression for meta variables and principal components
}
