% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvals.R
\name{plot_pvals}
\alias{plot_pvals}
\title{Plot  p-values}
\usage{
plot_pvals(pvals, name = NA, width = 8, height = 7)
}
\arguments{
\item{pvals}{A matrix of p-values.}

\item{name}{Name of the plot file.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}
}
\description{
Plot p-values
}
