% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval2z.R
\name{pval2z}
\alias{pval2z}
\title{Convert p-values to Z-scores}
\usage{
pval2z(pval, direction)
}
\arguments{
\item{pval}{Vector of two-sided test p-values.}

\item{direction}{Numeric or character vector of same length of pvalue vector. Examples of numeric valuses are log fold-change,
regression slope, or correlation coefficient. For charecter values, only "Up" and "Down" are allowed.}
}
\value{
Vector of Z-scores.
}
\description{
Counvert two-sided test p-values to Z-scores using \code{qnorm}.
}
