% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_fisher_test.R
\name{venn_fisher_test}
\alias{venn_fisher_test}
\title{Fisher exact test for 2-way Venn}
\usage{
venn_fisher_test(mat.sig, directional = TRUE)
}
\arguments{
\item{mat.sig}{Numeric matrix with two columns whose elements are {1, -1, 0} for significantly up, down, or
insignificant genes (rows) in a test (column) if \code{directional=TRUE}.
If \code{directional=FALSE}, we only consider significant and non significant genes, so elements are {0, 1}.}

\item{directional}{Logical; are we considering whether significant genes are up or down?
Otherwise, we only consider if genes are significant or not.}
}
\value{
Data frame with statistics from Fisher exact test.
}
\description{
Fisher exact test for enrichment of overlap in 2-way Venn.
}
